+++ packages/xrpl/src/models/transactions/index.ts
@@ -86,3 +86,5 @@ export {
   XChainModifyBridgeFlags,
   XChainModifyBridgeFlagsInterface,
 } from './XChainModifyBridge'
+export { Batch } from './batch'
+export { LedgerStateFix } from './ledgerStateFix'